Change name of APT installation section
Ubuntu may be the most common APT-using system, but not the original and certainly not the only. Also, the prominence of the distro may change with time. Better to allow n00bs to think "Is my distro a Debian derivative" and find the answer by web search rather than thinking "My distro is not Ubuntu, Fedora, or OSX...wat do?!".